Implement auto radius for all mass objects
