+++ core/src/net/sf/openrocket/rocketcomponent/Parachute.java
@@ -15,7 +15,6 @@ public class Parachute extends RecoveryDevice {
 	private double diameter;
 	private final double InitialPackedLength = this.length;
 	private final double InitialPackedRadius = this.radius;
-	private boolean autoRadius;
 
 	private Material lineMaterial;
 	private int lineCount = 6;
@@ -145,64 +144,6 @@ public class Parachute extends RecoveryDevice {
 	public String getComponentName() {
 		//// Parachute
 		return trans.get("Parachute.Parachute");
-	}
-
-	@Override
-	public void setRadius(double radius) {
-		radius = Math.max(radius, 0);
-
-		for (RocketComponent listener : configListeners) {
-			if (listener instanceof MassObject) {
-				((MassObject) listener).setRadius(radius);
-			}
-		}
-
-		if (MathUtil.equals(this.radius, radius) && (!autoRadius))
-			return;
-
-		this.autoRadius = false;
-		this.radius = radius;
-		fireComponentChangeEvent(ComponentChangeEvent.MASS_CHANGE);
-	}
-
-	@Override
-	public double getRadius() {
-		if (autoRadius) {
-			if (parent == null) {
-				return radius;
-			}
-			if (parent instanceof NoseCone) {
-				return ((NoseCone) parent).getAftRadius();
-			} else if (parent instanceof Transition) {
-				double foreRadius = ((Transition) parent).getForeRadius();
-				double aftRadius = ((Transition) parent).getAftRadius();
-				return (Math.max(foreRadius, aftRadius));
-			} else if (parent instanceof BodyComponent) {
-				return ((BodyComponent) parent).getInnerRadius();
-			} else if (parent instanceof RingComponent) {
-				return ((RingComponent) parent).getInnerRadius();
-			}
-		}
-		return radius;
-	}
-
-	public boolean isRadiusAutomatic() {
-		return autoRadius;
-	}
-
-	public void setRadiusAutomatic(boolean auto) {
-		for (RocketComponent listener : configListeners) {
-			if (listener instanceof Parachute) {
-				((Parachute) listener).setRadiusAutomatic(auto);
-			}
-		}
-
-		if (autoRadius == auto)
-			return;
-
-		autoRadius = auto;
-
-		fireComponentChangeEvent(ComponentChangeEvent.BOTH_CHANGE);
 	}
 	
 	@Override
